moved out of src, tested on test PyPI
