+++ pygame_gridcalculator/shapefactory.py
