+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pygame-gridcalculator
-version = 2.0.0
+version = 2.0.0a
 author = Dave Harding
 author_email = dave@punkamania.org
 description = For generating a virtual grid to map pixels against and generate shapes from for a pygame display.
@@ -15,11 +15,8 @@ classifiers =
     Development Status :: 5 - Production/Stable
 
 [options]
-package_dir=
-    =src
 packages = find:
 python_requires = >=3.8
 include_package_data = False
-install_requires = [
-    "pygame>=2.1.2"
-]
+install_requires =
+    pygame>=2.1.2
